more shapes

+++ compile.py
@@ -1,4 +1,6 @@
 import csv
+import os.path, time
+
 from jinja2 import Environment, FileSystemLoader, select_autoescape
 from glob import glob
 from os.path import basename
@@ -47,14 +49,6 @@ def enchantments(item_id):
     ]) + ']'
 
 
-def square_grid(block, limit):
-    for x in range(-limit, limit):
-        for y in range(-limit, limit):
-            for h in range(-limit, limit):
-                if (x % 2) > 0 and (y % 2) > 0 and (z % 2) > 0:
-                    yield f"setblock ~{x} ~{y} ~{z} {block}"
-
-
 def cylinder(radius, top, bottom=-1, walls='netherite_block'):
     x0 = 0
     y0 = 0
@@ -90,7 +84,7 @@ def cylinder(radius, top, bottom=-1, walls='netherite_block'):
         yield f"fill ~{x0 - y} ~{bottom} ~{y0 - x} ~{x0 - y} ~{top} ~{y0 - x} {walls}"
 
 
-def disk(radius, bottom=-1, floor='birch_planks', ring=None):
+def disk(radius, bottom=-1, floor='birch_planks', top=None, ring=None):
     x0 = 0
     y0 = 0
 
@@ -100,10 +94,12 @@ def disk(radius, bottom=-1, floor='birch_planks', ring=None):
     x = 0
     y = radius
 
-    yield f"fill ~{x0} ~{bottom} ~{y0 + radius} ~{x0} ~{bottom} ~{y0 + radius} {floor}"
-    yield f"fill ~{x0} ~{bottom} ~{y0 - radius} ~{x0} ~{bottom} ~{y0 - radius} {floor}"
-    yield f"fill ~{x0 + radius} ~{bottom} ~{y0} ~{x0 + radius} ~{bottom} ~{y0} {floor}"
-    yield f"fill ~{x0 - radius} ~{bottom} ~{y0} ~{x0 - radius} ~{bottom} ~{y0} {floor}"
+    top = top or bottom
+
+    yield f"fill ~{x0} ~{bottom} ~{y0 + radius} ~{x0} ~{top} ~{y0 + radius} {floor}"
+    yield f"fill ~{x0} ~{bottom} ~{y0 - radius} ~{x0} ~{top} ~{y0 - radius} {floor}"
+    yield f"fill ~{x0 + radius} ~{bottom} ~{y0} ~{x0 + radius} ~{top} ~{y0} {floor}"
+    yield f"fill ~{x0 - radius} ~{bottom} ~{y0} ~{x0 - radius} ~{top} ~{y0} {floor}"
 
     while x < y:
         if f >= 0:
@@ -114,8 +110,8 @@ def disk(radius, bottom=-1, floor='birch_planks', ring=None):
         ddf_x += 2
         f += ddf_x
         # floor
-        yield f"fill ~{x0 - x} ~{bottom} ~{y0 - y} ~{x0 + x} ~{bottom} ~{y0 + y} {floor}"
-        yield f"fill ~{x0 - y} ~{bottom} ~{y0 - x} ~{x0 + y} ~{bottom} ~{y0 + x} {floor}"
+        yield f"fill ~{x0 - x} ~{bottom} ~{y0 - y} ~{x0 + x} ~{top} ~{y0 + y} {floor}"
+        yield f"fill ~{x0 - y} ~{bottom} ~{y0 - x} ~{x0 + y} ~{top} ~{y0 + x} {floor}"
 
     if ring:
         yield from disk(radius-ring, bottom=bottom, floor='air')
@@ -128,9 +124,11 @@ def run(srcs=TEMPLATE_PATH, dst='data/uberop/functions'):
     for src in pathlib.Path(srcs).glob(join('**', '*.jinja')):
         final_src = str(src).replace(srcs_slash, '')
         final_dst = str(join(dst, final_src)).replace('.jinja', '.mcfunction')
-        print(f'compiling {final_src} to {final_dst}')
-        with open(final_dst, 'w') as f:
-            f.write(render(final_src))
+        # check file dates
+        if not os.path.isfile(final_dst) or os.path.getmtime(src) > os.path.getmtime(final_dst):
+            print(f'compiling {final_src} to {final_dst}')
+            with open(final_dst, 'w') as f:
+                f.write(render(final_src))
 
 
 if __name__ == "__main__":
